construção dos chunks/splittext e criaçao da nova feat para embedding com FAISS vectorStore
