+++ utils/embbedings.py
@@ -1,0 +1,2 @@
+from langchain.embeddings.openai import OpenAIEmbeddings
+from langchain.vectorstores import FAISS
