--- utils/text.py
@@ -1,9 +1,0 @@
-from PyPDF2 import PdfReader
-def process_files(files):
-    
-    text = ""
-    for file in files:
-        pdf_reader = PdfReader(file)
-        for page in pdf_reader.pages:
-            text += page.extract_text()
-    return text
